updated code

+++ 03 conditionals-loops/Areaofcircle.java
@@ -4,7 +4,7 @@ public class Areaofcircle {
     public static void main(String[] args) {
         Scanner in = new Scanner(System.in);
         System.out.print("Enter radius of circle: ");
-        int r = in.nextInt();
+        double r = in.nextDouble();
         double a = 3.14 * r * r;
         System.out.println("Area of the Circle is: "+a);
     }
